Replaced PreChaosNodeStatusCheck, PostChaosActiveNodeCountCheck, and getActiveNodeCount functions to nodes.go; updated the gcp-vm-instance-stop experiment for the same.
